% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dire_plot.R
\name{dire_plot}
\alias{dire_plot}
\title{dire plot Helper function that draws a bar chart when passed dota2 id
shows your winning rate in team dire}
\usage{
dire_plot(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
dire plot Helper function that draws a bar chart when passed dota2 id
shows your winning rate in team dire
}
\examples{
dire_plot(135474549)



}
