% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mmr.R
\name{get_mmr}
\alias{get_mmr}
\title{get_mmr function returns a dataframe with your nickname
 and matchmaking rate score when passed dota2 id}
\usage{
get_mmr(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
get_mmr function returns a dataframe with your nickname
 and matchmaking rate score when passed dota2 id
}
\examples{
get_mmr(135474549)



}
