% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmmr.R
\name{getmmr}
\alias{getmmr}
\title{getmmr function returns a dataframe with your nickname and matchmaking rate score when passed dota2 id}
\usage{
getmmr(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
getmmr function returns a dataframe with your nickname and matchmaking rate score when passed dota2 id
}
\examples{
getmmr(135474549)



}
