% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hero_rank.R
\name{hero_rank}
\alias{hero_rank}
\title{This function tells your top-rank heroes when passed dota2 id}
\usage{
hero_rank(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
This function tells your top-rank heroes when passed dota2 id
}
\examples{
hero_rank(135474549)



}
