% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_describe.R
\name{match_describe}
\alias{match_describe}
\title{Helper function that shows information about a specficial when passed match id}
\usage{
match_describe(matchID)
}
\arguments{
\item{matchID}{a dota2 match ID}
}
\description{
Helper function that shows information about a specficial when passed match id
}
\examples{
match_describe(4172245371)



}
