% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_details.R
\name{match_details}
\alias{match_details}
\title{This function returns a data frame shows player information about a specifical match when passed a match id}
\usage{
match_details(matchID)
}
\arguments{
\item{matchID}{a dota2 match ID}
}
\description{
This function returns a data frame shows player information about a specifical match when passed a match id
}
\examples{
match_details(4172245371)



}
