% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_wins.R
\name{match_wins}
\alias{match_wins}
\title{Helper function that shows winning information about a specficial match when passed match id}
\usage{
match_wins(matchID)
}
\arguments{
\item{matchID}{a dota2 match ID}
}
\description{
Helper function that shows winning information about a specficial match when passed match id
}
\examples{
match_wins(4172245371)


}
