% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiant_plot.R
\name{radiant_plot}
\alias{radiant_plot}
\title{radiant plot Helper function that draws a bar chart when passed dota2 id
shows your winning rate in team radiant}
\usage{
radiant_plot(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
radiant plot Helper function that draws a bar chart when passed dota2 id
shows your winning rate in team radiant
}
\examples{
radiant_plot(135474549)



}
