% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_lose_plot.R
\name{win_lose_plot}
\alias{win_lose_plot}
\title{win_lose_plot Helper function that draws a bar chart when passed dota2 id}
\usage{
win_lose_plot(dota2ID)
}
\arguments{
\item{dota2ID}{your dota2 account ID}
}
\description{
win_lose_plot Helper function that draws a bar chart when passed dota2 id
}
\examples{
win_lose_plot(135474549)



}
